+++ src/autorest-core/lib/configuration.ts
@@ -27,6 +27,7 @@ import { BlameTree } from './source-map/blaming';
 import { MergeOverwriteOrAppend, resolveRValue } from './source-map/merging';
 import { TryDecodeEnhancedPositionFromName } from './source-map/source-map';
 import { safeEval } from './ref/safe-eval';
+import { OutstandingTaskAwaiter } from "./outstanding-task-awaiter"
 
 const untildify: (path: string) => string = require("untildify");
 
@@ -240,7 +241,7 @@ export class ConfigurationView {
       this.config = this.rawConfig;
     }
     this.suppressor = new Suppressor(this);
-    this.Message({ Channel: Channel.Debug, Text: `Creating ConfigurationView : ${configs.length} sections.` });
+    // this.Message({ Channel: Channel.Debug, Text: `Creating ConfigurationView : ${configs.length} sections.` });
   }
 
   public get Keys(): Array<string> {
@@ -393,17 +394,18 @@ export class ConfigurationView {
           let blameTree: BlameTree | null = null;
 
           try {
+            const originalPath = JSON.stringify(s.Position.path);
             while (blameTree === null) {
               try {
                 blameTree = this.DataStore.Blame(s.document, s.Position);
               } catch (e) {
                 const path = s.Position.path as string[];
                 if (path) {
-                  this.Message({
-                    Channel: Channel.Verbose,
-                    Text: `Could not find the exact path ${JSON.stringify(path)} for ${JSON.stringify(m.Text)}`
-                  });
                   if (path.length === 0) {
+                    this.Message({
+                      Channel: Channel.Verbose,
+                      Text: `Could not find the exact path ${originalPath} for ${JSON.stringify(m.Text)}`
+                    })
                     throw e;
                   }
                   path.pop();
@@ -738,13 +740,20 @@ export class Configuration {
 
     return createView().Indexer;
   }
-
   public static async DetectConfigurationFile(fileSystem: IFileSystem, configFileOrFolderUri: string | null, messageEmitter?: MessageEmitter, walkUpFolders: boolean = false): Promise<string | null> {
+    const files = await this.DetectConfigurationFiles(fileSystem, configFileOrFolderUri, messageEmitter, walkUpFolders);
+
+    return From<string>(files).FirstOrDefault(each => each.toLowerCase().endsWith("/" + Constants.DefaultConfiguration)) ||
+      From<string>(files).OrderBy(each => each.length).FirstOrDefault() || null;
+  }
+
+  public static async DetectConfigurationFiles(fileSystem: IFileSystem, configFileOrFolderUri: string | null, messageEmitter?: MessageEmitter, walkUpFolders: boolean = false): Promise<Array<string>> {
+    const results = new Array<string>();
     const originalConfigFileOrFolderUri = configFileOrFolderUri;
 
-    //
-    if (!configFileOrFolderUri || configFileOrFolderUri.endsWith(".md")) {
-      return configFileOrFolderUri;
+    // null means null!
+    if (!configFileOrFolderUri) {
+      return results;
     }
 
     // try querying the Uri directly
@@ -752,50 +761,40 @@ export class Configuration {
       try {
         const content = await fileSystem.ReadFile(configFileOrFolderUri);
         if (content.indexOf(Constants.MagicString) > -1) {
-          return configFileOrFolderUri;
+          // the file name was passed in!
+          return [configFileOrFolderUri];
         }
       } catch (e) {
         // that didn't work... try next
       }
     }
 
-    // search for a config file, walking up the folder tree
-    while (configFileOrFolderUri !== null) {
-      // scan the filesystem items for the configuration.
-      const configFiles = new Map<string, string>();
 
-      for (const name of await fileSystem.EnumerateFileUris(EnsureIsFolderUri(configFileOrFolderUri))) {
-        if (name.endsWith(".md")) {
-          const content = await fileSystem.ReadFile(name);
-          if (content.indexOf(Constants.MagicString) > -1) {
-            configFiles.set(name, content);
-          }
+    // scan the filesystem items for configurations.
+    for (const name of await fileSystem.EnumerateFileUris(EnsureIsFolderUri(configFileOrFolderUri))) {
+      if (name.endsWith(".md")) {
+        const content = await fileSystem.ReadFile(name);
+        if (content.indexOf(Constants.MagicString) > -1) {
+          results.push(name);
         }
       }
+    }
 
-      if (configFiles.size > 0) {
-        // it's the readme.md or the shortest filename.
-        const found =
-          From<string>(configFiles.keys()).FirstOrDefault(each => each.toLowerCase().endsWith("/" + Constants.DefaultConfiguration)) ||
-          From<string>(configFiles.keys()).OrderBy(each => each.length).First();
-
-        return found;
-      }
-
+    if (walkUpFolders) {
       // walk up
       const newUriToConfigFileOrWorkingFolder = ResolveUri(configFileOrFolderUri, "..");
-      configFileOrFolderUri = !walkUpFolders || newUriToConfigFileOrWorkingFolder === configFileOrFolderUri
-        ? null
-        : newUriToConfigFileOrWorkingFolder;
+      if (newUriToConfigFileOrWorkingFolder !== configFileOrFolderUri) {
+        results.push(... await this.DetectConfigurationFiles(fileSystem, newUriToConfigFileOrWorkingFolder, messageEmitter, walkUpFolders))
+      }
+    } else {
+      if (messageEmitter && results.length === 0) {
+        messageEmitter.Message.Dispatch({
+          Channel: Channel.Warning,
+          Text: `No configuration found at '${originalConfigFileOrFolderUri}'.`
+        });
+      }
     }
 
-    if (messageEmitter) {
-      messageEmitter.Message.Dispatch({
-        Channel: Channel.Warning,
-        Text: `No configuration found at '${originalConfigFileOrFolderUri}'.`
-      });
-    }
-
-    return null;
+    return results;
   }
 }
